requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
pydantic-settings
pytest
pytest-mock
httpx
